Do not update conditions until SmartNIC starts reporting health (#11997)
In current CMD code, SmartNIC is declared HEALTHY as soon as it
is admitted to the cluster. However, on the scale setup we have
observed cases in which the SmartNIC takes time to perform
post-Admission tasks and succesfully report health, causes
spurious healthy -> unhealthy -> healthy  transitions.

With this change, CMD ignores conditions in the admission
request and starts updating them in ApiServer only when the
SmartNIC object succesfully sends the first update.