linux_ionic: clean up sparse complaints (#12043)
The Linux sparse compiler/typechecker is very handy for
hunting down the byte-swapping issues in the driver so we
have a chance at working in big-endian machines.  I haven't
tested this in such a machine, so it may not do all the job,
but it gets the most obvious things worked out.

The first thing is to fix all the DMA accesses to be sure they
are byteswapped correctly.  The more tricky part is fixing
up the dev_cmd accesses through the PCI bar, as some of our
fields have special requirements.

In order to be sure we get data in and out of the dev_cmd space,
we need to use the ioreadXX and iowriteXX functions to be sure
that each architecture's different access methods are handled.

These functions do byteswapping for us, so we need to be
sure we don't swap them again.  However, where we are moving
a block of values in and out of the bar, we need to be sure
the swapping happens correctly.  In our case, this sometimes
means we need byteswap ahead of time and then do a byte-by-byte
copy rather than word-by-word, so that we don't get a pair of
u16's or a string of u8's swapped in a really odd way.

There are changes here to the ionic_if.h file, but they are all
type changes which shouldn't be disruptive to the other drivers.

Signed-off-by: Shannon Nelson <snelson@pensando.io>