Force SmartNIC health updates propagation to ApiServer (#12116)
SmartNIC status updates consisting only of Conditions field changes
were not propagated from CMD to ApiServer when the reference object
in CMD had empty Conditions field, because empty is the zero value
for the field. With this change, the Conditions field is always
compared and if reference is empty and update non-empty, the update
is propagated.