An adminq timeout (in this case it was on a crypto key update command) can happen during FW upgrade (#12190)
when nicmgr reaches the IONIC_FW_INSTALL stage. This would block nicmgr for many seconds and delay
its adminq response. The keyword here is 'delay' as eventually nicmgr would be able to send a response.
(Note: Neel says he will fix this blocking behavior when he gets around to it.)

However, an adminq timeout means the sonic driver would give up on the command and delete its completion
context. Any delayed response would lead to a crash because the driver would be working on
context memory that's already gone.

The fix is, whenever there is a timeout, nullify the context callback in the affected adminq descriptor so
a delayed response would simply be dropped.

In addition, sonic driver will switch to a 60 second timeout similar to the freebsd ionic driver.