ionic: use the right lif adminq for macvlan (#12808)
When doing adminq requests in the macvlan offload lif, use the
slave lif's adminq, not the master lif's adminq.  In this case,
we were working out the doorbell page information before we
had learned the lif's hw_index.  Moving the calculations to
after the call to ionic_dev_cmd_lif_init() fixes the problem.

Signed-off-by: Shannon Nelson <snelson@pensando.io>