Simplify txrx napi (#13005)
* linux_ionic: inline the interrupt handling utilities

Most of the interrupt handling got inlined earlier in Allen's
intr patch, but here are a couple more routines that can be
inlined to help the compiler give us better code in the RxTx
speed path.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: simplify tx flush

Pull the disparate functions together into a single Tx flush
function in one file so the compiler can better optimize
the code.  This replaces the use of the more generic q and
cq processing used by the AdminQ and the NotifyQ - they can
deal with the overhead of a generalized queue processing,
but the Tx and Rx fastpaths need all the tuning they can get.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: simplify the rx clean handling

Pull the disparate functions together into a single Rx clean
function in one file so the compiler can better optimize
the code.  This replaces the use of the more generic q and
cq processing used by the AdminQ and the NotifyQ - they can
deal with the overhead of a generalized queue processing,
but the Tx and Rx fastpaths need all the tuning they can get.

Signed-off-by: Shannon Nelson <snelson@pensando.io>