add retry to ephemeral overlay txn commit failures (#13487)
Address VS-430.

Txn commit fails due to parallel updates to the object. In case of the
VS-430 the failure is due to NPM updating the App object status. The
Delete operation from user fails because the version changes in between
building the Txn object and commit due to NPM updating the App object.
Add a retry for TxnCommit failures to address such cases.