Add support for Devlink and for ethtool -d (#13602)
* linux_ionic: add ethtool register dump

Print out the contents of dev_info and dev_cmd registers in
the ethtool -d output.  This will best be useful with an
addition to the ethtool userland tool to decode and pretty-print
the data.  Until then, we get a hex dump to hand decode.  Yay.

This was requested by our upstream friends in order to
replace the register dump in debugfs.  We'll keep the debugfs
dec_cmd and dev_info output in our out of tree driver for now.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: add devlink interface

Add a devlink interface into the driver for access to
information that isn't normally available through
ethtool or the iplink interface.

This is by request from our upstream friends, partly in order
to replace some of our debugfs use.

Example:
	$ ./devlink -j -p dev info pci/0000:b6:00.0
	{
	    "info": {
		"pci/0000:b6:00.0": {
		    "driver": "ionic",
		    "serial_number": "FLM18420073",
		    "versions": {
			"fixed": {
			    "fw_version": "0.11.0-50",
			    "fw_status": "0x1",
			    "fw_heartbeat": "0x716ce",
			    "asic_type": "0x0",
			    "asic_rev": "0x0"
			}
		    }
		}
	    }
	}

Signed-off-by: Shannon Nelson <snelson@pensando.io>