Upstream strip (#13988)
* linux_ionic: separate adminq cmd from api

The RDMA feature is not yet supported in the upstream
driver.  In order to make it easier to compare and support
the upstream driver, move the adminq command interface out
of the api files.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: always use debug stats

There's really no need to make the debug stats selectable
as they will always be wanted.  Debug stats are no help
in debugging a problem if they haven't been compiled in.
Also add some of the detailed stats only seen in debugfs to
the stats reported in ethtool -S.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: clean up debugfs

Our upstream friends pointed out several things that can get
cleaned up in our debugfs use.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: clean up a bunch of random bits

There were a lot of random nits picked by our upstream friends,
this patch addresses several of them.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: use devlink to create ionic

Use devlink_alloc() and devlink_free() to manage the memory space
for the device specific ionic struct.  We also set up a couple of
aliases for when working in older OSs without devlink support.

Signed-off-by: Shannon Nelson <snelson@pensando.io>