Prevent resolver from blocking election (#14633)
Change in leader results in an notification to the Service Tracker which
in turn updates the service instances. This tries to service watchers by
pushing to a channel. IF the watch handler blocks then subsequent Leader
change notifications get blocked and leads to a bad cluster state.

This PR prevents this by wrapping the watcher Send() action in a timeout.