linux_ionic: fix up some coalesce value issues (#14774)
Set the initial coalesce value to something other than 0 in order
to be sure that some coalescing is happening by default.  For now,
we'll use the same value that the FreeBSD driver is using, and in
the future we may want to do some tuning.

We apply this value only to the Rx queues since the Tx queues are
sharing the Rx interrupts.  Also, fix the ethtool set_coalesce so
we don't try applying a new value from the user if the netdev is
down and no queues/intrs are allocated yet.

Signed-off-by: Shannon Nelson <snelson@pensando.io>