Fix a deadlock in CMD (#14953)
When CMD receives an update for a Host object, it calls UpdateNICPairingStatus
while processing. UpdateNICPairingStatus in turn may call UpdateHostPairingStatus.
UpdateHostPairingStatus may try to acquire a lock for the Host object again,
causing a deadlock.
The fix consists in decoupling UpdateNICPairingStatus and UpdateHostPairingStatus
and making sure that we don't try to reacquire the lock on the same Host object.