Workaround for a Kubernetes/etcd/grpc issue (#15105)
Kubernetes apiserver crashes repeatedly if it cannot connect to etcd
running on node1. This is due to the fact that etcd client library
uses a version of the gRPC load-balancer which always checks for
the hostname of the first server in the list even when trying
to connect to subsequent server.

Once the bug is fixed in gRPC and the fix ported in Kubernets apiserver
container we can get rid of this workaround

Issue: https://github.com/kubernetes/kubernetes/issues/72102

Also increases the number of retained CMD log files