ionic: AdminQ enhancements, split eth vs. rdma (#15676)
* ionic: Split adminq paths between ethernet and external callers

Internal ethernet callers will continue to use the adminq. These accesses
are protected by the LIF lock to allow the LIF to be reset cleanly.

External callers (RDMA) will use an API that funnels the commands to the
dev_cmd interface. This allows RDMA to keep working even during an
ethernet LIF reset.

* ionic: Recover from missed AdminQ interrupts [PS-1911]

QEMU/KVM can lose an interrupt if it is asked to reconfigure interrupts
while one is arriving.

This could affect any interrupt, but adminq is the only one I've seen
actually happen.