Sync code with upstream (#15952)
* linux_ionic: add ionic prefix to linux specific structs and variables

The upstream Linux reviewers required us to add the "ionic_" prefix
to all our structs and constants.  This patch gets all the Linux
driver specific structs and constants.  We'll do the ionic_if.h
pieces later.

Some of these name changes force related code changes, including
how some of the function variables are declared and initialized.
Much of this comes from a demand by the netdev maintainer that
all function variable declarations be formated in "reverse
Christmas tree" format.  It seems a little picky, and forces
some odd rework, but it is what is expected upstream.  Doing it
here makes my life easier when comparing our internal driver
with the Linux in-tree driver.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: ignore debugfs errors

The message from our upstream friends is to ignore errors out
of the debugfs facility.  We've already done much of that
conversion, but here are a few more bits that got missed.

The upstream driver code removes even more, but that's because
the debugfs service routines in upstream have recently been
worked over to be much more tolerant of bad pointers and other
errors.  Since this code needs to work with older kernels as
well, we still need to do some of the internal error checking.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: more bits and pieces from upstream

These are a couple more little tweaks to get our internal code
closer to the upstream driver code.

Signed-off-by: Shannon Nelson <snelson@pensando.io>