Implement Eventq support in the Linux driver (#16219)
* linux_ionic: fixup core formatting issues

A few checkpatch complaints have cropped up recently, so
fix them up before moving on to the rest of the patchset.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: add error checking to macvlan offload setup

There were a couple of API calls that weren't being checked,
but if they failed, crashes can happen later.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: change param from lif to ionic

Use an ionic pointer rather than the lif in ionic_intr_alloc()
and ionic_intr_free() so that in the future we can allocate
interrupts before we have a lif.

Originally from Allen's draft code.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: convert lif list to radix_tree

In the near future we will need to be able to use the lif index
to find a lif.  Converting to radix_tree means we won't have to
search the list, and yet we won't need to allocate the whole
pointer array when we're usually only going to need a handful
of slots.

In our upstream driver the Linux folks will want us to use
the newer xarrays rather than radix_tree, which is fair,
but isn't back-ported to kernels in the older distro's that
we need to support.  This code may change in the future to
make that different of opinion easier to manage, likely
through a new chunk of kcompat code.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: replace nslaves with nlifs

Let's keep track of the number of lifs we are using rather
than one of the derivitives of the number of lifs.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: rename rdma eqs field

Before adding fields to track the Eth Event Queues, we
need to rename the RDMA Event Queues fields so as to
prevent any confusion as to which set we're working with.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: add basic eq supoprt

In configurations where we want lots of lifs for macvlan offload
support, and lots of queues in each lif, we can easily run out
of interrupt resources and have to find ways to limit our usage,
either by using fewer queues and/or lifs, or by multiplexing
the interrupt needs.

This patch adds a new EventQueue feature that can be used for
multiplexing the interrupts if we find that we can't get enough
from the system to support our configuration.  We can create a
small number of EQs that use interrupts, and have the TxRx queue
pairs subscribe to event messages that come through the EQs.
The interrupt messages normally generated by packet events
become events in the EQ, and the EQ handler will schedule a
napi session for the queue that has an event.

The use of EQs is driven by the firmware configuration - if
it shows support for ethernet event queues (eq_count in the
device identity), the driver will use them instead of regular
interrupts per TxRx queuepair.

At the moment, AdminQ does not use EventQueue, but could be
added later.  A better choice might be to have the slave lifs
use the lif0 AdminQ rather than creating individual AQs for
every single lif.

Originally from Allen's draft code.

Signed-off-by: Shannon Nelson <snelson@pensando.io>