ionic_rdma: set eq_count before creating admin cqs (#16469)
Fixes out-of-bounds array access of eq_vec, if there are more admin cq
than eq.

Issue (kernel panic) was seen on a host with few cpus, which limited the
num eqs created.  We might also consider limiting the number of adminq
in to no more than the number of cpus, as well.