linux_ionic: set netdev default name (#17537)
If the host system's udev fails to set a new name for the
network port, it remains as eth0 or eth1 and there is no
NETDEV_CHANGENAME netlink event to trigger the driver to
send the name down to Naples.  Since multiple calls to set
the lif name don't seem to matter, let's add a call early on
to set the default netdev name to be sure we have something
set in Naples.  Then when udev gets around to changing it
we can update it to teh actual name the system will be using.

This will fix some iota tests that are failing on RHEL8 as
the udev seems a little broken at the moment.

Signed-off-by: Shannon Nelson <snelson@pensando.io>