linux_ionic: detect bad eq setup from older FW (#17656)
There was a change in the FW API after 1.1.x and the Lif Identity info
for eth_eq_count returns as garbage with the older FW, and the q_identify
calls are not supported.  This patch cleans up some error handling so
that this doesn't cause errors in the driver startup.  Now if the eq
fails the driver falls back to regular Tx Rx queue interrupt usage, and
if the driver sees the q_identify opcode fail on the older FW it will be
able to continue on.

Signed-off-by: Shannon Nelson <snelson@pensando.io>