Linux driver support for fw upgrade (#17854)
* linux_ionic: add fw_status to devlink output

Add fw_status to the devlink output for monitoring
when desired without relying on sysfs.

devlink dev info pci/0000:b5:00.0

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: tx and rx queues state follows link state

When link goes down, tear down the Tx and Tx queues.  When
link comes back up, rebuild the queues.  This lets us release
memory resources for devices that aren't going to use them,
and prepares us for support of fw upgrade.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: keep old lif dentry

Don't replace the lif->dentry until we know we have
a good new value.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: protect ethtool functions from fw reset

If we've torn down much of the driver infrastructure for
fw update, we need to protect ethtool operations from
stubbing its toe on the missing pieces.  Here we add a
new bitflag IONIC_LIF_F_FW_RESET which will be true when
the FW is going through a reset cycle.  This will replace
the current IONIC_LIF_F_FW_READY in order to change the
sense of some of the existing FW reset logic.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: ignore eexist on rx filter add

Don't worry if the rx filter add firmware request fails on
EEXIST, at least we know the filter is there.  Same for
the delete request, at least we know it isn't there.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: pull lif-init-queues into lif-init

Keep the two chunks of lif initialization code together
so we can use the function later for handling fw reset.

It was split out a couple of months ago in order to add
event-queues support, so here we keep event-queue support
in the middle of the lif-init, but only if we're init'ing
the master lif.

Also, move the ionic_lif_init_queues() function up in the
file so we can get rid of the forward declaration.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: remove lifs on fw reset

When the FW RESET event comes to the driver from the firmware,
tear down the LIFs.  This event signals a FW upgrade where we
need to quiesce all operations and wait for the FW to restart.
The FW will continue the update process once it sees all the
LIFs are reset.  When the update process is done it will set
the fw_status back to RUNNING.  Meanwhile, the heartbeat
check continues and when the fw_status is seen as set to
running we can rebuild the LIFs.

We expect that there was a LINK_DOWN event before this to stop
the queues, and there will be a LINK_UP event afterwards to
get things started again.

Note that this changes the name and reverses the sense of the
related flag such that it is to be set only when the FW is in
RESET mode.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: check for link when link down

While the link is down, have the heartbeat watchdog also
check the link for changes in case no Event is sent.

Meanwhile, sometimes the link status check runs into itself
or the ionic_open() thread, and can either confuse itself,
throw incorrect error messages, or even miss the link message.
This patch cleans up the code and gets the right messages out
to the log file when they happen.

 - rearrange the link_up/link_down messages so that we announce
   link up when we first notice that the link is up when the
   driver loads, and decouple the link_up/link_down messages
   from the UP and DOWN netdev state.
 - add error checking of ionic_qcq_disable() so as to not have
   to wait too long on shutdown when everything it timeout out.
 - add a state to indicate transitions between netdev UP and
   DOWN so as to prevent recursion.
 - don't clean the dev_cmd when we notice that the fw has halted,
   it's possible it might still execute the request.

Signed-off-by: Shannon Nelson <snelson@pensando.io>