Error case cleanup around fw-upgrade (#18799)
* linux ionic: init of rxtx needs to wait for link up

In order to keep clean the model of rx and tx queues only
built when link is up, we need to not create the queues at
open time unless link is already seen as up.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: fix crash from rmmod during fw-upgrade

When the driver is waiting in fw-upgrade for the FW to restart,
it has removed the rxqcqs[] txqcqs[] arrays.  When rmmod goes
through the ionic_remove(), it ends up in ionic_qcqs_free()
which tries to remove them again and crashes when trying to
dereference a NULL pointer.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: use a short timeout on shutdown

If the FW is broken when the user has called to remove the driver,
the driver can end up waiting the full 30 second timeout on the
first FW command, and this wait is multiplied by the number of
PCI devices supplied by the NIC with the broken FW.  Since the
rmmod is holding the rtnl_lock that whole time, the Linux kernel
gets real grumpy and starts throwing softlock error messages.

Here we fix the devcmd_timeout to be a smaller value when in
shutdown so as to minimize the hang problem.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: no vf operations while fw in reset

Block VF operations when the FW is in a RESET state, such
as when doing a FW upgrade.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: remove irq affinity before irq_free

There is an error path where the irq affinity is set at qcq
alloc, but the qcq doesn't make it up to enable, so doesn't
go through disable where the affinity is removed.  In this
case the kernel throws a WARN when we free the interrupt
int ionic_qcq_free().  Here we add a little call to be
sure the affinity is cleaned up.

Signed-off-by: Shannon Nelson <snelson@pensando.io>