Fix a bogus eq count from older FW (#18890)
* linux ionic: keep ionic dev on lif init fail

If the basic ionic interface works but the lif creation fails, 
don't fail the probe.  This will allow us to use the driver to
help inspect the hw/fw/pci interface for debugging purposes.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: protect from old fw neqs garbage

Older FW that doesn't know about EventQueues leaves some
garbage around in the device identity response that can make
eth_eq_count show a huge bogus number, which then screws with
the resource calculations in ionic_lif_size().  The easiest
way to protect from this is to clamp the eth_eq_count to a
maximum valid value.  This might make older FW look like it
has support for EQs, but when the driver tries to set them
up the FW will NAK the qtype and the driver will continue on
using Tx/Rx queue interrupts instead as it normally would.

Signed-off-by: Shannon Nelson <snelson@pensando.io>