iota boot_naples_v2: do sync after login, fwupdate (#19411)
The host reboot would be started before the completion of the naples fw
update command.  The expect would match a command prompt from before
naples fw update command was run, and proceed immediately to rebooting
the host before waiting for the command to complete.

This inserts a semi-unique string to expect, after login, and after
fwupate -s commands, to ensure that the naples command line is
synchronized at those points in boot_naples_v2.