Avoid inline updates to cluster object after leadership change (#19478)
When there is a leadership change, the new leader needs to bring up
the Kubernetes control plane as soon as possible, so that other
nodes can register and the pods can be re-distributed.
Updating the cluster object takes time because ApiServer may not be
available, etcd instances may be busy, etc. so it is better to do
it asynchronously (on a timer) and not inline.