rxhash disable, numa aware irq, code cleanup (#19531)
* linux ionic: fold ionic_lif_init_queues back into ionic_lif_init

This code was pulled into a separate functino while ago when
EventQueue support was first brought in.  Since then some changes
and refactoring have happened such that this is no longer needed.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: support ethtool rxhash disable

We can disable rxhashing by setting rss_types to 0.  The user
can toggle this with "ethtool -K <ethX> rxhash off|on",
which calls into the .ndo_set_features callback with the
NETIF_F_RXHASH feature bit set or cleared.  This patch adds
a check for that bit and updates Naples FW if necessary.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: improve irq numa locality

Spreading the interrupts across the CPU cores is good for load
balancing, but not necessarily as good when using a CPU/core
that is not part of the NUMA local CPU.  If it can be localized,
the kernel's cpumask_local_spread() service will pick a core
that is on the node close to the PCI device.

The queue structs are allocated with devm_kzalloc() which
already is putting them local to the device's NUMA node.
This might affect performance a little if we ever end up with
interrupts on a distant node, but should be fine for now.

Signed-off-by: Shannon Nelson <snelson@pensando.io>