Increase num sges allowed per RDMA packet (#19353) (#19599)
Increase num sges from 4 to 12 for one RDMA packet to fetch payload,
if SGEs point to physically contiguous memory, as in dma_mr
region with reserved lkey. Currently 3 DMA cmds are reserved for 1 SGE,
with a total of 12 for 4 SGEs. For reserved lkey case, 1 DMA cmd is
sufficient (as memory is physically contiguous) and hence 1 RDMA packet can be
allowed to span across 12 SGES