PCI connection report and error (#19586)
* linux ionic: print pci bus lane info

Print the PCI bus lane information so people can keep an
eye out for poor configurations that might not support
the advertised speed.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: catch rebooted fw

When the Naples FW reboots, the PCI link goes down and the
fw_status and fw_heartbeat stop.  Shortly there after the
fw_status is seen as up and the driver tries to restart, which
fails and can lead to a host OS crash.  This comes from an
issue in the way that PR-18958 changed the check for setting
work->fw_status in ionic_headtbeat_check().  This does not
affect the similar PR-18960 in the 1.3-E.

Signed-off-by: Shannon Nelson <snelson@pensando.io>