Double commit to master (#18197)
* Don't enable tx/rx queues when port is down

* Add vmnic name to some logging

* Fix a race condition when reading link states

* Improve the logic we have in heartbeat detection
Currently, when we see failures in adminq cmds twice, we consider
the device is in bad state and stop posting any further adminq cmds.
This is not complete correct  since we may see errors regarding on
duplicated mac filter apply. In this case, device is not
in bad state. We should only consider device is in bad state when
timeout failures are captured twice consecutively.

* Store ionic_keep_posting as per device(PF) variable
Currently, if we have two Naples per host, if one Naples
has timeout issue, we will block the adminq posting for
the other Naples. This is not correct. We need to treat
ionic_keep_posting as per device variable, we only skip
adminq commands come from the device that has failed twice
consecutively due to timeout.

* Return better status to the caller when timeout happens

* Chnage the type of keep_posting_cmds from vmk_atomic8 to vmk_atomic64
vmk_atomic8/16/32 are not supported in ESXi 6.5 DDK. Will use vmk_atomic64
for all versions for now.

* Address an issue when timeout happens during rapid link-flap for rss queue

* Replace queue not available log with a queue busy counter

* Add padding for incoming rx packets that less than 60 bytes
Arp reply coming from int_mnic0 is less than 60 bytes and got
dropped by vSwitch. In order to avoid this drop, we need to pad
this kind of packet.

* Driver shouldn't set uplink_handle->is_started if tx/rx queues are not enabled