ionic_eth, ionic_rdma: Upstreaming work, rd 7 (#20155)
* ionic_eth: Fix up build for kernel 5.5.x

* ionic_rdma: Add compat for linux 5.5 and 5.6

* ionic_rdma: Switch from module parameters to configfs on recent kernels

We cannot upstream any new module parameters.

* ionic_rdma: Truthiness in advertising wrt max_sge

The Naples DSC supports 16 fragments per WR, but there is another
per-packet limit that we are not able to effectively advertise to
clients. This per-packet limit is 12 in the send path and 4 in the
read and receive paths.

NetApp has requested that we advertise 12 rather than 16 for the
send path. BSD does not have max_recv_sge, so we are effectively
advertising 12 for receive even though the per-packet limit is 4.
This will be addressed through release notes.

The advertised value of max_sge_rd is left at 4 since we do not
expect anyone to bump up against this limit.

All of these values assume that each individual SGE fragment is
physically contiguous.

* ionic_rdma: Add compat for MOFED 4.7

* ionic_eth, ionic_rdma: Update docker for Linux 5.5 and MOFED 4.7

* ionic_rdma: Use defined(HAVE_X) form to reduce nesting

* ionic_eth: Trivial whitespace cleanups

Some of this trailing whitespace shows up in the diff from 1.2 to 1.4.
I would rather have to explain a diff that removes ws than one that adds ws.

* ionic_eth, ionic_rdma: Add NETAPP_PATCH tags

These diffs trip us up every time we import new files into ONTAP.
Code them here so that the files will build in both places.

NetApp has a __FreeBSD_version, but it's out of sync with upstream.
Also, let's not have another daylong firedrill about CDP/LLDP/VLAN.