ionic: remove pragma packed (#19839)
The strong suggestion from Linux upstream folks is to remove
the use of "#pragma packed" and replace it with tags on each
struct/union definition that actually needs it.  This is safer
on the various compilers and architectures.

The __packed tag on structs and unions signals to the compiler
that no padding is needed and that some fields accessed
unaligned.  Since we specifically added padding in the
definitions, and tried to keep fields on the right boundaries,
most of these structs don't need the tag.

This change to the ionic_if.h file should have no consequences
in the use of the definitions in the drivers that use this file.
The resulting new code should have no compatibilty issues with
existing drivers and firmware.

Signed-off-by: Shannon Nelson <snelson@pensando.io>