Brel bugs (#20833)
* linux_ionic: limit max Tx descriptors

Asking for 16K size Tx rings fails with kernel messages:
  ionic 0000:b5:00.0 enp181s0: Cannot allocate queue DMA memory
  ionic 0000:b5:00.0: qcq alloc of tx0 failed -12
With 16k ring size this ends up trying to reserve a DMA block of
over 4Mbyte for the Tx cmd, comp, and sg descriptor space for
each ring.

By limiting the Tx rings to 8k we can keep the per-ring descriptor
DMA block to a little over 2Mbyte.

Fixes: PS-2523
Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: no link msg on queue parameter change

There's no need for a link up message after a queue parameter
change such as mtu or descriptor ring size.  The Link Up msg
is a result of turning off carrier in ionic_lif_stop() which
is done to prevent bogus Tx Timeout msgs.  Let's turn it back
on if it was on in order to prevent unnecessary Link messages.

The netif_tx_disable() does what netif_tx_stop_all_queues() tries
to do, but does it better, so we can remove it.

Also, add a couple of parameter changed messages.

Fixes: PS-2532

Signed-off-by: Shannon Nelson <snelson@pensando.io>