Move the CPU LIF programming to FTE. There is occasionally a scenario (#22250)
where CPU LIF gets programmed earlier than PMD init and due to this they
go out of sync. Moving CPU LIF programming to FTE is not very ideal
because if we end up supporting more FTE cores then we might have to
move it to a place after all the FTEs are inited.