linux_ionic: Upstream rework for v5.6 (#22053)
* linux_ionic: some fixes from upstream

After doing a bunch of work on the upstream driver, I found
several things in our local driver that need little tweaks.
This is a collection of small fixes to make our out-of-tree
driver closer to what we have upstream.  Most of these are
fairly innocuous, and some are minor bug fixes found when
working through the upstream code..

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: print default xcvr info

Even if the transceiver is not recognized, let ethtool print
the xcvr information to help with debugging.  Also, use the
SFP symbols to match upstream code.  We define the SFP symbols
ourselves here in order to be able to compile on older Linux
distributions that don't have the sfp.h include file.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* ionic: set station addr only if needed

The code was working too hard and in some cases was trying to
delete filters that weren't there, generating a potentially
misleading error message.
    IONIC_CMD_RX_FILTER_DEL (32) failed: IONIC_RC_ENOENT (-2)

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: move debugfs add-delete to match alloc-free

Move the qcq debugfs add to the queue alloc, and likewise move
the debugfs delete to the queue free.  The LIF debugfs add
also needs to be moved, but the del is already in the LIF free.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: move irq request to qcq alloc

Move the irq request and free out of the qcq_init and deinit
and into the alloc and free routines where they belong for
better resource management.  This helps keep all the alloc
work separate from the init work so that we can make use of
the separate layers.

The use of netif_napi_add() and netif_napi_del() in the init
phase seems a bit confusing, but they actually are init and
de-init functions in the stack by adding to and removing the
napi context from the napi processing list.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: enhanced tx timeout handling

Well, only slightly enhanced by printing out the tx queue id
that newer Linux kernels make available to us.  In the future
we can use this to do a specific queue reset rather than
blasting the whole lif.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: check for qcq exist before delete

Add some checking for existence before trying to delete the
queue structures.  This is most useful in some failure recovery
where something might not have gotten fully set up before we
had to abandon ship.

In the enable code, we swap the tx and rx so taht rx is first
because it is the rx that has the interrupt handling and this
assures that all is ready for the tx.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: fw upgrade rework

Our upstream Linux friends weren't impressed with the first
implementation of fw-upgrade support and suggested a few
changes.  Those have now been accepted upstream, so this is
to get us back in sync in the internal driver.

The design difference here is that originally we were calling
ionic_open() and ionic_close() when link when up and down
in order to create and delete the queues.  The argument from
upstream was that once we had the queue resources we should not
let them go and then try to re-alloc them again.  This rework
separates the queue enable from the queue alloc and init, and
we simply disable and re-enable the queues on link down and
link up, keeping the existing allocated buffers, descriptors,
and interrupts.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: scnprintf is safer than snprintf

This is directly from an upstream patch:
Since snprintf() returns the would-be-output size instead of the
actual output size, the succeeding calls may go beyond the given
buffer limit.  Fix it by replacing with scnprintf().

Ref: 38e0f746c456 ("net: ionic: Use scnprintf() for avoiding potential buffer overflow")
Signed-off-by: Shannon Nelson <snelson@pensando.io>

* ionic: add hw error and drop stats to ethtool stats
    
Gather a few more hardware stats from the DSC to be
displayed in the ethtool stats output.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: update filter id after replay

When we replay the rx filters after a fw-upgrade we get new
filter_id values from the FW, which we need to save and update
in our local filter list.  This allows us to delete the filters
with the correct filter_id when we're done.

Also fix up some of the locking and debug messages.

Signed-off-by: Shannon Nelson <snelson@pensando.io>